+++ src/ncsdis.cpp
@@ -28,7 +28,8 @@
 
 #include <vector>
 
-#include "src/common/version.h"
+#include "src/version/version.h"
+
 #include "src/common/ustring.h"
 #include "src/common/util.h"
 #include "src/common/error.h"
