+++ src/unerf.cpp
@@ -28,7 +28,8 @@
 #include <cstring>
 #include <cstdio>
 
-#include "src/common/version.h"
+#include "src/version/version.h"
+
 #include "src/common/ustring.h"
 #include "src/common/strutil.h"
 #include "src/common/error.h"
