+++ src/unkeybif.cpp
@@ -27,7 +27,8 @@
 #include <list>
 #include <vector>
 
-#include "src/common/version.h"
+#include "src/version/version.h"
+
 #include "src/common/util.h"
 #include "src/common/ustring.h"
 #include "src/common/error.h"
