+++ src/unnds.cpp
@@ -25,7 +25,8 @@
 #include <cstring>
 #include <cstdio>
 
-#include "src/common/version.h"
+#include "src/version/version.h"
+
 #include "src/common/ustring.h"
 #include "src/common/error.h"
 #include "src/common/platform.h"
